mt8195: l2c: Turn off L3C SRAM and reconfigure as L3 cache
On mt8195, we use half of the L3 cache as SRAM before DRAM is ready.
After DRAM is ready, L3 is still used as SRAM which causes hang issues
in scenarios such as suspend-to-ram S3.
Mediatek SoC uses part of the L3 cache as SRAM before DRAM is ready.
After DRAM is ready, we should invoke disable_l2c_sram to reconfigure
the L3C SRAM as L3 cache.

Signed-off-by: Ryan Cho <ryan.cho@mediatek.com>
Signed-off-by: Amjad Ouled-Ameur <aouledameur@baylibre.com>
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>