GENIO: fix(mt8365): incorrect vcore voltage after resume
During suspend, SPM will set the Vcore to the lowest level.
After resume, the DVFS should set the Vcore to the appropriate level.
Since Yocto does not support DVFS, Vcore cannot be set to the correct
voltage after resume. So set the Vcore voltage to 0.8V in SPM resume
to solve this problem.

Change-Id: I5e493946786c2615140562d40a9737f7668b1982
Signed-off-by: Chris-QJ Chen <chris-qj.chen@mediatek.com>