GENIO: mediatek: 8188: apusys: set maximum frequency by platform
Before, for Genio-700, we set maximum frequency: 900 mhz for apusys
mdla during the initialization.

For now, we have new platform: Genio-510, it also uses the same apusys
but running mdla in different frequencies and voltages. Due to the
maximum frequency of apusys in Genio-510 is different to Genio-700,
so we have to set correct maximum frequency for apusys mdla according
to different platform.

Currently we determine platform by segment id read from efuse and set
correct maximum frequency for apusys mdla to support Genio-700 and
Genio-510 platforms.

Change-Id: Iab0cbd683529adb2383deb68cdefbc258b05abb9
Signed-off-by: kidd-kw.chen <kidd-kw.chen@mediatek.com>