GENIO: plat/mediatek: iommu: Modify the order of IOMMU_ATF_CMD
[Problem]
The order of SMC CMD IOMMU_ATF_CMD is the same as ALPS project,
because this patch is porting form other project which is using
ALPS kernel and original TFA-v2.6. But the order is mismatch with
original linux kernel and this will make some SMC CMD does work.

[Solution]
Attach extra CMD for secure needs after linux kernel upstream CMD.

Change-Id: I014d801523986ea70ae0eb69b46ca8d0170ad5f5
Signed-off-by: Chengci.Xu <chengci.xu@mediatek.com>