GENIO: mediatek: 8188: apusys: Set ENABLE_SW_BUCK_CTL to 0
The SW_BUCK_CTL is known to cause power-on issues on some Genio 700
boards. To mitigate this issue, we set the power supplies of APU
to "always-on". This brings a cost of higher system idle power.

In order to reduce the system idle power, disable the ENABLE_SW_BUCK_CTL
in both trusted-firmware-a and APUSYS driver. This configures the
APU firmware to control the power supplies directly without involving
the kernel driver.

This has a side effect that the reference numbers in
/sys/kernel/debug/regulator/regulator_summary no longer reflects the
actual regulator states, which is been controlled by APU firmware
directly.

One important note is that both trust-firmware-a and APUSYS must
have the same ENABLE_SW_BUCK_CTL configuration, and the kernel's
device tree must be configured to set APU power supplies to always-on.
Failing to do so leads to critical APUSYS failues.

CR-Id: ALPS09375096
Change-Id: I613e9014a12cb748ae5731119e83b78f6f1ba69c
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>